fix(items): use responsive tab overflow on item sheets
A GM's spell sheet renders five tabs (Spell, Manifestation, Mercurial,
Description, Judge-Only Description), which wrapped over the sheet header.
Extract the actor sheet's responsive-tabs overflow handling into a shared
module/responsive-tabs.mjs, use it from both sheets, and switch the item
tab partial to the responsive markup — overflowing tabs tuck into the
ellipsis dropdown. The shared setup now also returns a cleanup that
disconnects the ResizeObserver and document listener, and both sheets tear
down the previous render's observer before wiring a new one.

Verified live: an actor-owned spell shows four tabs plus the ellipsis, and
the Judge-Only Description tab opens from the dropdown.
